user_program zupt cal: hide the individual configs, only go through auto calibrate
it doesn't make sense to set individual zupt configs since it depends on vehicle setup. Should only set by auto calibration.
Hide the "set manually" option and don't show the individual config values
show "Calibrated" if any are nonzero, or "Not calibrated" if all 0 (as it will be on a new unit or after reset).